requirements.txt
pytest-homeassistant-custom-component==0.13.304
ruff==0.14.10
bump2version==1.0.1